Generalized MMC class, migrating server.
