+++ adv_package/loader/models/__init__.py
@@ -1,2 +1,3 @@
-from .wideresnet import WideResNet, WideResNetMMCL
+from .wideresnet import WideResNet
 from .fulldenoiser import FullDenoiser
+from .mmc_model import create_mmc_object
